Code cleanup + bugfixes.
- Tidied up main event loop.
- Moved click logic into normal mode.
- Fixed grid mode scrolling.
- Added check for activation key conflicts before forking.
- Removed oneshot mode.
- Changed default scroll keys to minimize conflicts.

+++ src/hint_drw.c
@@ -32,7 +32,7 @@
 static Display *dpy;
 
 static struct hint *hints;
-static int nhints;
+static size_t nhints;
 static int border_radius;
 
 static Pixmap label_pixmap;
